requirements.txt
beautifulsoup4>=4.12.3
langchain>=0.0.27
lxml>=5.3.0
numpy>=2.1.2
openai>=1.53.0
pdfplumber>=0.11.4
pillow>=10.4.0
plotly>=5.24.1
psycopg2-binary>=2.9.10
pymupdf>=1.24.13
pypdf2>=3.0.1
python-dotenv>=1.0.0
streamlit>=1.39.0
tenacity>=9.0.0
tiktoken>=0.8.0
requests>=2.31.0  # Added for Ollama API calls
cryptography>=42.0.0
